header traits
