+++ ssdp-rs/src/message/header.rs
@@ -2,7 +2,7 @@
 //! - [UpnpHeader] for storing & finding values
 //! - enums / structs for all standard header fields with standard key name, parsing & display logic
 
-use std::collections::HashMap;
+use std::{collections::HashMap, fmt::Display};
 
 use crate::message::ParseError;
 
@@ -36,3 +36,57 @@ impl<'h> UpnpHeader<'h> {
         self.0.get(key).copied()
     }
 }
+
+/// Marker trait for Upnp header fields, with details of the relevant key
+pub trait Header {
+    /// Key as per spec
+    const HEADER_KEY: &'static str;
+}
+
+/// Handles constructing valid header lines.
+///
+/// This is a separate trait from [Header] to allow for it to also be implemented on `Option<H>`
+pub trait HeaderExt {
+    /// Generate a valid header line
+    fn to_header(&self) -> String;
+
+    /// Write a valid header line to `f` including new-line
+    fn write_header(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result;
+}
+
+impl<H: Header + Display> HeaderExt for H {
+    fn to_header(&self) -> String {
+        format!("{}: {}", Self::HEADER_KEY, self)
+    }
+
+    fn write_header(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        writeln!(f, "{}", self.to_header())
+    }
+}
+
+impl<H: Header + HeaderExt> HeaderExt for Option<H> {
+    /// Generate a valid header line
+    ///
+    /// #### Note
+    /// - This will output an empty `String` for `None`.
+    ///   If this is not what you want consider using `.map(|h| h.to_header())` which
+    ///   will give you an `Option<String>` instead.
+    fn to_header(&self) -> String {
+        match self {
+            Some(header) => header.to_header(),
+            None => String::new(),
+        }
+    }
+
+    /// Write a valid header line to `f` including new-line
+    ///
+    /// #### Note
+    /// - `None` entries are handled nicely (no-op) *without* generating a blank line
+    fn write_header(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            Some(header) => header.write_header(f),
+            None => Ok(()),
+        }
+    }
+}
+
