+++ ssdp-rs/src/message/mod.rs
@@ -24,7 +24,7 @@ mod error;
 mod header;
 
 pub use error::ParseError;
-pub use header::UpnpHeader;
+pub use header::{Header, HeaderExt, UpnpHeader};
 
 use crate::{MULTICAST, SSDP_PORT};
 const UPNP_VERSION: &str = "2.0";
@@ -767,58 +767,6 @@ impl Header for UpnpPort {
     const HEADER_KEY: &'static str = "SEARCHPORT.UPNP.ORG";
 }
 
-/// Marker trait for Upnp header fields, with details of the relevant key
-trait Header {
-    /// Key as per spec
-    const HEADER_KEY: &'static str;
-}
-
-/// Handles constructing valid header lines.
-///
-/// This is a separate trait from [Header] to allow for it to also be implemented on `Option<H>`
-trait HeaderExt {
-    /// Generate a valid header line
-    fn to_header(&self) -> String;
-
-    /// Write a valid header line to `f` including new-line
-    fn write_header(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result;
-}
-
-impl<H: Header + Display> HeaderExt for H {
-    fn to_header(&self) -> String {
-        format!("{}: {}", Self::HEADER_KEY, self)
-    }
-
-    fn write_header(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        writeln!(f, "{}", self.to_header())
-    }
-}
-
-impl<H: Header + HeaderExt> HeaderExt for Option<H> {
-    /// Generate a valid header line
-    ///
-    /// #### Note
-    /// - This will output an empty `String` for `None`.
-    ///   If this is not what you want consider using `.map(|h| h.to_header())` which
-    ///   will give you an `Option<String>` instead.
-    fn to_header(&self) -> String {
-        match self {
-            Some(header) => header.to_header(),
-            None => String::new(),
-        }
-    }
-
-    /// Write a valid header line to `f` including new-line
-    ///
-    /// #### Note
-    /// - `None` entries are handled nicely (no-op) *without* generating a blank line
-    fn write_header(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Some(header) => header.write_header(f),
-            None => Ok(()),
-        }
-    }
-}
 #[derive(Debug, Clone, PartialEq)]
 pub struct Notification {
     location: Option<String>,
